worked on query logic
It should be able to run a query on the server, return the json file, and parse it so it is ready to read from in javascript. I included an example query that I had tested. I tested this on the old version of the webpage and had it working just fine. I expect it to work here too.